celery working but losing tracks
