+++ sales_agent_service/src/worker/celery_app.py
@@ -1,10 +1,10 @@
-
 from celery import Celery
 
 celery_app = Celery(
     "worker",
     broker="redis://localhost:6379/0",    # Redis as message broker
-    backend="redis://localhost:6379/0"      # Redis as result backend
+    backend="redis://localhost:6379/0",      # Redis as result backend
+    include=['src.worker.tasks']  # Add this line to include tasks
 )
 
 celery_app.conf.update(
